Move message parsing and recipient calculation into separate modules
