+++ rust/main/chains/hyperlane-sealevel/src/provider.rs
@@ -1,13 +1,10 @@
-use std::collections::HashSet;
 use std::sync::Arc;
 
 use async_trait::async_trait;
-use lazy_static::lazy_static;
 use solana_sdk::signature::Signature;
 use solana_transaction_status::{
     option_serializer::OptionSerializer, EncodedTransaction, EncodedTransactionWithStatusMeta,
-    UiInstruction, UiMessage, UiParsedInstruction, UiParsedMessage, UiTransaction,
-    UiTransactionStatusMeta,
+    UiTransaction, UiTransactionStatusMeta,
 };
 use tracing::warn;
 
@@ -18,21 +15,13 @@ use hyperlane_core::{
 };
 
 use crate::error::HyperlaneSealevelError;
+use crate::provider::message_parser::parsed_message;
+use crate::provider::recipient::RecipientProvider;
 use crate::utils::{decode_h256, decode_h512, decode_pubkey};
 use crate::{ConnectionConf, SealevelRpcClient};
 
-lazy_static! {
-    static ref NATIVE_PROGRAMS: HashSet<String> = HashSet::from([
-        solana_sdk::bpf_loader_upgradeable::ID.to_string(),
-        solana_sdk::compute_budget::ID.to_string(),
-        solana_sdk::config::program::ID.to_string(),
-        solana_sdk::ed25519_program::ID.to_string(),
-        solana_sdk::secp256k1_program::ID.to_string(),
-        solana_sdk::stake::program::ID.to_string(),
-        solana_sdk::system_program::ID.to_string(),
-        solana_sdk::vote::program::ID.to_string(),
-    ]);
-}
+mod message_parser;
+mod recipient;
 
 /// A wrapper around a Sealevel provider to get generic blockchain information.
 #[derive(Debug)]
@@ -81,7 +70,7 @@ impl SealevelProvider {
     }
 
     fn sender(hash: &H512, txn: &UiTransaction) -> ChainResult<H256> {
-        let message = Self::parsed_message(txn)?;
+        let message = parsed_message(txn)?;
 
         let signer = message
             .account_keys
@@ -90,41 +79,6 @@ impl SealevelProvider {
         let pubkey = decode_pubkey(&signer.pubkey)?;
         let sender = H256::from_slice(&pubkey.to_bytes());
         Ok(sender)
-    }
-
-    fn recipient(hash: &H512, txn: &UiTransaction) -> ChainResult<H256> {
-        let message = Self::parsed_message(txn)?;
-
-        let programs = message
-            .instructions
-            .iter()
-            .filter_map(|ii| {
-                if let UiInstruction::Parsed(iii) = ii {
-                    Some(iii)
-                } else {
-                    None
-                }
-            })
-            .map(|ii| match ii {
-                UiParsedInstruction::Parsed(iii) => &iii.program_id,
-                UiParsedInstruction::PartiallyDecoded(iii) => &iii.program_id,
-            })
-            .filter(|program_id| !NATIVE_PROGRAMS.contains(*program_id))
-            .collect::<Vec<&String>>();
-
-        if programs.len() > 1 {
-            Err(HyperlaneSealevelError::TooManyNonNativePrograms(Box::new(
-                *hash,
-            )))?;
-        }
-
-        let program_id = programs
-            .first()
-            .ok_or(HyperlaneSealevelError::NoNonNativePrograms(Box::new(*hash)))?;
-
-        let pubkey = decode_pubkey(program_id)?;
-        let recipient = H256::from_slice(&pubkey.to_bytes());
-        Ok(recipient)
     }
 
     fn gas(meta: &UiTransactionStatusMeta) -> ChainResult<U256> {
@@ -154,15 +108,6 @@ impl SealevelProvider {
             .as_ref()
             .ok_or(HyperlaneSealevelError::EmptyMetadata)?;
         Ok(meta)
-    }
-
-    fn parsed_message(txn: &UiTransaction) -> ChainResult<&UiParsedMessage> {
-        Ok(match &txn.message {
-            UiMessage::Parsed(m) => m,
-            m => Err(Into::<ChainCommunicationError>::into(
-                HyperlaneSealevelError::UnsupportedMessageEncoding(Box::new(m.clone())),
-            ))?,
-        })
     }
 
     async fn block_info_by_height(&self, slot: u64) -> Result<BlockInfo, ChainCommunicationError> {
@@ -225,7 +170,7 @@ impl HyperlaneProvider for SealevelProvider {
 
         Self::validate_transaction(hash, txn)?;
         let sender = Self::sender(hash, txn)?;
-        let recipient = Self::recipient(hash, txn)?;
+        let recipient = RecipientProvider::recipient(hash, txn)?;
         let meta = Self::meta(txn_with_meta)?;
         let gas_used = Self::gas(meta)?;
         let fee = self.fee(meta)?;
